+++ changes.markdown
@@ -1,5 +1,10 @@
 Changelog
 =========
+### 2014-02-03
+* Update to Elasticsearch 1.0.0.RC2
+* Update GeohashGrid aggregation to reflect name change
+* Update aggregations tests to compensate for response refactoring
+
 ### 2014-01-29
 * Update to Elasticsearch 1.0.0.RC1. This is a breaking change.
 * Add aggregations